backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
azure-ai-inference==1.0.0b3
azure-search-documents==11.4.0
azure-identity==1.15.0
openai==1.10.0
httpx==0.26.0
python-multipart==0.0.6
pillow==10.2.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
